building&test running fixes
