+++ jbpm-process-svg/pom.xml
@@ -2,12 +2,11 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <parent>
-    <groupId>org.drools</groupId>
-    <artifactId>droolsjbpm-integration</artifactId>
-    <version>7.0.0-SNAPSHOT</version>
+    <groupId>org.jbpm</groupId>
+    <artifactId>jbpm</artifactId>
+    <version>6.2.0.Final-Center-1</version>
   </parent>
 
-  <groupId>org.kie</groupId>
   <artifactId>jbpm-process-svg</artifactId>
   <name>jBPM Process SVG Manipulation</name>
 
@@ -15,17 +14,10 @@
     <dependency>
       <groupId>batik</groupId>
       <artifactId>batik-svg-dom</artifactId>
-      <exclusions>
-        <exclusion>
-          <artifactId>xercesImpl</artifactId>
-          <groupId>xerces</groupId>
-        </exclusion>
-      </exclusions>
     </dependency>
     <dependency>
       <groupId>batik</groupId>
       <artifactId>batik-util</artifactId>
     </dependency>
   </dependencies>
-
 </project>
